requirements.txt
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Jinja2==2.10
MarkupSafe==1.0
SQLAlchemy==1.2.0b3
WTForms==2.1
WebOb==1.7.3
Werkzeug==0.12.2
bcrypt==3.1.4
beautifulsoup4==4.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
docopt==0.6.2
dominate==2.3.1
fanstatic==0.16
gunicorn==19.7.1
idna==2.6
igdb-api-python==1.0
igdbapi==0.2
itsdangerous==0.24
oauthlib==2.0.6
pip==9.0.1
pipreqs==0.4.9
pycparser==2.18
requests==2.18.4
requests-oauthlib==0.8.0
setuptools==36.8.0
six==1.11.0
tweepy==3.5.0
ujson==1.35
urllib3==1.22
users==1.0.dev0
virtualenv==15.1.0
visitor==0.1.3
wtf==0.1